Understand how the arrayToBow works, tested it, and moved it to its own module.
