+++ test/main.spec.js
@@ -1,8 +1,20 @@
 const expect = require("chai").expect;
 const getEmoteFromMessage = require("../helpers/recording").getEmoteFromMessage;
 const sliceByTime = require("../helpers/recording").sliceByTime;
+const arrayToBow = require("../helpers/recording").arrayToBow;
 
 describe("Emote Ticker", () => {
+  var stickerData = new Map();
+  beforeEach(() => {
+    // Set up a mock sticker map. This may make no sense to you 
+    // if you're new dev. Learn what ES2016 Maps are:
+    // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Map
+    stickerData.set(1559363923054, '376849f2c000955_274351');
+    stickerData.set(1559363923799, '376849f2c000955_274351');
+    stickerData.set(1559363928385, '36095b58a009292_250342');
+    stickerData.set(1559363928957, '36095b58a009292_250342');
+  })
+
   it("Lets me know if Mocha's working", () => {
     expect(false).to.be.false;
   });
@@ -21,14 +33,6 @@ describe("Emote Ticker", () => {
   })
 
   it("Lists the emotes used within a certain timeframe", () => {
-    stickerData = new Map();
-    // Set up a mock sticker map. This may make no sense to you 
-    // if you're new dev. Learn what ES2016 Maps are:
-    // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Map
-    stickerData.set(1559363923054, '376849f2c000955_274351');
-    stickerData.set(1559363923799, '376849f2c000955_274351');
-    stickerData.set(1559363928385, '36095b58a009292_250342');
-    stickerData.set(1559363928957, '36095b58a009292_250342');
     // Call sliceByTime with a start time less than the earliest emote
     // and an end time later than the latest emote. Then add the 
     // emote (sticker) usage and their time stamps. 
@@ -40,9 +44,16 @@ describe("Emote Ticker", () => {
         '36095b58a009292_250342'
     ]
     )
+  })
 
-
-
+  it("Tallies the number of times an emote is used for a list of emotes", () => {
+    const slicedStickers = sliceByTime(1559363923053, 1559363928958, stickerData);
+    expect(arrayToBow(slicedStickers)).to.eql(
+      {
+        '376849f2c000955_274351': 2,
+        '36095b58a009292_250342': 2
+      }
+    );
   })
 
   xit("Stores sticker info into a db", () => {
